requirements-cloud.txt
influxdb_client
pandas
pyyaml
requests
pyarrow
google-cloud-storage